# clinic-api/docs/swagger.yaml
basePath: /api
definitions:
  clinic.Address:
    properties:
      city:
        example: São Paulo
        type: string
      state:
        example: SP
        type: string
      street:
        example: Av. Paulista, 1000
        type: string
      zip_code:
        example: 01310-100
        type: string
    type: object
  clinic.Banking:
    properties:
      account:
        type: string
      agency:
        example: "1234"
        type: string
      bank:
        example: Banco do Brasil
        type: string
    type: object
  clinic.ClinicStatus:
    enum:
    - pending
    - active
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusActive
  clinic.CreateInput:
    properties:
      address:
        $ref: '#/definitions/clinic.Address'
      banking:
        $ref: '#/definitions/clinic.Banking'
      description:
        example: Clínica odontológica completa
        type: string
      document:
        description: Document is the CPF/CNPJ. Immutable after creation.
        example: "12345678000199"
        type: string
      email:
        example: contato@clinicasorriso.com.br
        type: string
      legal_name:
        example: Clínica Sorriso LTDA
        type: string
      opening_hours:
        example: Seg a Sex, 8h às 18h
        type: string
      phone:
        example: (11) 4000-1000
        type: string
      specialties:
        example:
        - Ortodontia
        - Implantodontia
        items:
          type: string
        type: array
      trade_name:
        example: Clínica Sorriso
        type: string
      website:
        example: www.clinicasorriso.com.br
        type: string
    type: object
  clinic.UpdateInput:
    properties:
      address:
        $ref: '#/definitions/clinic.Address'
      banking:
        $ref: '#/definitions/clinic.Banking'
      description:
        example: Clínica odontológica completa
        type: string
      document:
        example: "12345678000199"
        type: string
      email:
        example: contato@clinicasorriso.com.br
        type: string
      legal_name:
        example: Clínica Sorriso LTDA
        type: string
      opening_hours:
        example: Seg a Sex, 8h às 18h
        type: string
      phone:
        example: (11) 4000-1000
        type: string
      specialties:
        items:
          type: string
        type: array
      trade_name:
        example: Clínica Sorriso
        type: string
      website:
        example: www.clinicasorriso.com.br
        type: string
    type: object
  clinic.addressDTO:
    properties:
      city:
        example: São Paulo
        type: string
      state:
        example: SP
        type: string
      street:
        example: Av. Paulista, 1000
        type: string
      zip_code:
        example: 01310-100
        type: string
    type: object
  clinic.bankingResponse:
    properties:
      account:
        example: 56789-0
        type: string
      agency:
        example: "1234"
        type: string
      bank:
        example: Banco do Brasil
        type: string
    type: object
  clinic.clinicResponse:
    properties:
      address:
        $ref: '#/definitions/clinic.addressDTO'
      banking:
        $ref: '#/definitions/clinic.bankingResponse'
      created_at:
        type: string
      description:
        example: Clínica odontológica completa
        type: string
      document:
        example: "12345678000199"
        type: string
      email:
        example: contato@clinicasorriso.com.br
        type: string
      id:
        example: a0000000-0000-0000-0000-000000000001
        type: string
      legal_name:
        example: Clínica Sorriso LTDA
        type: string
      opening_hours:
        example: Seg a Sex, 8h às 18h
        type: string
      phone:
        example: (11) 4000-1000
        type: string
      specialties:
        items:
          type: string
        type: array
      status:
        allOf:
        - $ref: '#/definitions/clinic.ClinicStatus'
        enum:
        - pending
        - active
        example: active
      trade_name:
        example: Clínica Sorriso
        type: string
      updated_at:
        type: string
      website:
        example: www.clinicasorriso.com.br
        type: string
    type: object
  clinic.listResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/clinic.clinicResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  problem.Details:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      detail:
        example: request validation failed
        type: string
      errors:
        items:
          $ref: '#/definitions/problem.FieldError'
        type: array
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
      type:
        example: about:blank
        type: string
    type: object
  problem.FieldError:
    properties:
      detail:
        example: is required
        type: string
      field:
        example: clinic_id
        type: string
    type: object
info:
  contact: {}
  description: |-
    API de gestão de clínica odontológica: clínicas, dentistas e pagamentos via Pix (simulado).
    Todos os endpoints de negócio ficam sob o prefixo /api; o único fora dele é o health check em /health.
    Erros seguem o formato RFC 9457 (application/problem+json) em todo o corpo, com um campo "code" estável para tratamento programático além do "status" HTTP.
  title: Clinic API
  version: "1.0"
paths:
  /clinics:
    get:
      parameters:
      - default: 20
        description: Page size
        in: query
        maximum: 100
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Free-text search (trade name, legal name, specialties)
        in: query
        name: q
        type: string
      - description: Filter by city (substring, case-insensitive)
        in: query
        name: city
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clinic.listResponse'
      summary: List/search clinics (paginated)
      tags:
      - clinics
    post:
      consumes:
      - application/json
      description: Creates a clinic with status "pending". It becomes "active" automatically
        once it has at least one dentist marked as administrator and one marked as
        legal representative.
      parameters:
      - description: Clinic to create
        in: body
        name: clinic
        required: true
        schema:
          $ref: '#/definitions/clinic.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/clinic.clinicResponse'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/problem.Details'
        "409":
          description: document already belongs to another clinic
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Create a clinic
      tags:
      - clinics
...
  /clinics/{id}:
    delete:
      description: Marks the clinic as deleted (DeletedAt is set); it stops appearing
        in list/get and its document becomes reusable for a new clinic. Existing dentists
        and payments are not affected.
      parameters:
      - description: Clinic ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: clinic deleted
        "400":
          description: id is not a valid UUID
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic not found
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Soft-delete a clinic
      tags:
      - clinics
    get:
      parameters:
      - description: Clinic ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clinic.clinicResponse'
        "400":
          description: id is not a valid UUID
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic not found
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Get a clinic by id
      tags:
      - clinics
    put:
      consumes:
      - application/json
      description: Only non-null fields in the body are applied. "document" cannot
        be changed after creation and returns 400 DOCUMENT_IMMUTABLE if attempted.
      parameters:
      - description: Clinic ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: clinic
        required: true
        schema:
          $ref: '#/definitions/clinic.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clinic.clinicResponse'
        "400":
          description: validation error or attempt to change document
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic not found
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Update a clinic (partial)
      tags:
      - clinics
...
schemes:
- http
swagger: "2.0"